# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nemesis API",
    "version": "1.0.0",
    "description": "Artists platform API - Manage artworks, events, and artist profiles"
  },
  "servers": [
    {
      "url": "http://localhost:5005",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/artworks/{id}/video": {
      "post": {
        "tags": [
          "Artworks"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "content-length",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isPaid": {
                    "example": "any"
                  },
                  "duration": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/videos/{artworkId}/access": {
      "get": {
        "tags": [
          "Videos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "artworkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/videos/{artworkId}/stream": {
      "get": {
        "tags": [
          "Videos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "artworkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/videos/": {
      "get": {
        "tags": [
          "Videos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artistId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isPaid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/videos/{artworkId}/purchase": {
      "post": {
        "tags": [
          "Videos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "artworkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/videos/purchased": {
      "get": {
        "tags": [
          "Videos"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authToken",
        "description": "JWT token stored in HTTP-only cookie"
      }
    },
...
  }
}